short_scoring
Experimental branch to see if representing the per IP topic probabilities in 16 bits each saves us significant time and memory.

If this branch doesn't show the desired improvements... the only thing to do would be to go into spark lda and change it's representation of the probabilities, which is more bother than it's probably worth.

+++ spot-ml/src/main/scala/org/apache/spot/dns/model/DNSSuspiciousConnectsModel.scala
@@ -60,7 +60,7 @@ import scala.util.{Failure, Success, Try}
   * @param inEntropyCuts         Quantile cut-offs for discretizing the subdomain entropy in word construction.
   */
 class DNSSuspiciousConnectsModel(inTopicCount: Int,
-                                 inIpToTopicMix: Map[String, Array[Double]],
+                                 inIpToTopicMix: Map[String, Array[Short]], // SHORT testing
                                  inWordToPerTopicProb: Map[String, Array[Double]],
                                  inTimeCuts: Array[Double],
                                  inFrameLengthCuts: Array[Double],
@@ -290,8 +290,8 @@ object DNSSuspiciousConnectsModel {
       .rdd
       .map({ case (ipToTopicMixRow: Row) => ipToTopicMixRow.toSeq.toArray })
       .map({
-        case (ipToTopicMixSeq) => (ipToTopicMixSeq(0).asInstanceOf[String], ipToTopicMixSeq(1).asInstanceOf[Seq[Double]]
-          .toArray)
+        case (ipToTopicMixSeq) => (ipToTopicMixSeq(0).asInstanceOf[String], ipToTopicMixSeq(1).asInstanceOf[Seq[Short]]
+          .toArray)   // SHORT test
       })
       .collectAsMap
       .toMap
